src/app/video-chat/app/leave/__test__/FeedbackModal.test.tsx
import React from "react";
import { render, fireEvent, waitFor } from "@testing-library/react";
import FeedbackModal from "../modal/FeedbackModal";
import AppTheme from "../../../../../theme";

jest.mock("react-redux", () => ({
  useDispatch: jest.fn(),
  useSelector: jest.fn(),
}));
jest.mock("next/navigation", () => ({
  useRouter: jest.fn().mockReturnValue({
    push: jest.fn(),
  }),
}));

describe("FeedbackModal", () => {
  test("renders correctly", () => {
    const { getByText } = render(
      <AppTheme>
        <FeedbackModal open={true} onClose={() => {}} />
      </AppTheme>
    );

    expect(getByText("評価アンケート")).toBeInTheDocument();
    expect(getByText("Q1.")).toBeInTheDocument();
    expect(getByText("Q2.")).toBeInTheDocument();
  });

  test("handles rating selection", () => {
    const { getByText } = render(
      <AppTheme>
        <FeedbackModal open={true} onClose={() => {}} />
      </AppTheme>
    );

    fireEvent.click(getByText("1"));
    fireEvent.click(getByText("3"));
  });

  test("handles feedback submission", async () => {
    const { getByText, getByPlaceholderText } = render(
      <AppTheme>
        <FeedbackModal open={true} onClose={() => {}} />
      </AppTheme>
    );

    fireEvent.click(getByText("この内容で回答"));
    await waitFor(() =>
      expect(getByText("※ 必須項目が選択されていません。")).toBeInTheDocument()
    );

    fireEvent.click(getByText("1"));
    fireEvent.click(getByText("この内容で回答"));

    // You might want to mock the API call and test if it's called with the correct data
    // For brevity, I'm skipping the mocking part here
    // Asserting that the onClose callback is called
  });
});
